Extract login dialog into dedicated view
